# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "领域知识个性化生成与多智能体协同决策系统",
    "description": "领域知识个性化生成与多智能体协同决策系统后端API\n\n## 认证方式\n- 除 `/auth/login`、`/auth/register`、健康检查外，所有接口需在请求头携带 `Authorization: Bearer <access_token>`\n- Token 通过 `POST /api/v1/auth/login` 获取，有效期由 `JWT_EXPIRE_MINUTES` 控制\n- Token 过期后通过 `POST /api/v1/auth/refresh` 刷新\n\n## 统一响应格式\n所有接口返回 `{code, message, data, timestamp}` 结构，`code=200` 表示成功",
    "contact": {
      "name": "知识系统团队",
      "url": "https://github.com/knowledge-system"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "当前环境（相对路径）"
    },
    {
      "url": "http://localhost:8000/api/v1",
      "description": "开发环境"
    },
    {
      "url": "https://staging.example.com/api/v1",
      "description": "预发布环境"
    },
    {
      "url": "https://api.example.com/api/v1",
      "description": "生产环境"
    }
  ],
  "paths": {
...
    "/api/v1/knowledge/upload": {
      "post": {
        "tags": [
          "知识库"
        ],
        "summary": "上传知识库文档",
        "description": "上传知识库文档，自动解析切片并建立向量索引\n\n支持 multipart/form-data（文件上传）和 application/json（纯文本创建）",
        "operationId": "upload_document_api_v1_knowledge_upload_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "请求参数错误（文件格式不支持、内容为空等）"
          },
          "401": {
            "description": "未授权（Token缺失或过期）"
          },
          "403": {
            "description": "权限不足（需要管理员角色）"
          },
          "404": {
            "description": "资源不存在（文档ID无效）"
          },
          "413": {
            "description": "上传文件超出大小限制"
          },
          "422": {
            "description": "请求体验证失败"
          },
          "500": {
            "description": "服务器内部错误（解析失败、向量索引异常等）"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/knowledge/docs": {
      "get": {
        "tags": [
          "知识库"
        ],
        "summary": "获取文档列表",
        "description": "获取知识库文档列表，支持分页、搜索、筛选\n\n- **page**: 页码\n- **page_size**: 每页数量\n- **keyword**: 关键词搜索（标题/标签）\n- **industry**: 行业过滤\n- **status**: 状态过滤",
        "operationId": "get_doc_list_api_v1_knowledge_docs_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Page Size"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Keyword"
            }
          },
          {
            "name": "industry",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Industry"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/docs/{doc_id}": {
      "get": {
        "tags": [
          "知识库"
        ],
        "summary": "获取文档详情",
        "description": "根据ID获取文档详细信息",
        "operationId": "get_doc_detail_api_v1_knowledge_docs__doc_id__get",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Doc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "知识库"
        ],
        "summary": "更新文档信息",
        "description": "更新文档信息（标题、行业、标签等）",
        "operationId": "update_doc_api_v1_knowledge_docs__doc_id__put",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Doc Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeDocUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "知识库"
        ],
        "summary": "删除文档",
        "description": "删除文档（软删除）",
        "operationId": "delete_doc_api_v1_knowledge_docs__doc_id__delete",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Doc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/docs/batch-delete": {
      "post": {
        "tags": [
          "知识库"
        ],
        "summary": "批量删除文档",
        "description": "批量删除文档",
        "operationId": "batch_delete_docs_api_v1_knowledge_docs_batch_delete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeBatchDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/knowledge/search": {
      "post": {
        "tags": [
          "知识库"
        ],
        "summary": "知识库相似度检索",
        "description": "按知识点相似度检索知识库\n\n- **query**: 检索查询\n- **industry**: 行业过滤\n- **doc_id**: 指定文档\n- **top_k**: 返回结果数量\n- **min_similarity**: 最低相似度",
        "operationId": "search_knowledge_api_v1_knowledge_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "请求参数错误（文件格式不支持、内容为空等）"
          },
          "401": {
            "description": "未授权（Token缺失或过期）"
          },
          "403": {
            "description": "权限不足（需要管理员角色）"
          },
          "404": {
            "description": "资源不存在（文档ID无效）"
          },
          "413": {
            "description": "上传文件超出大小限制"
          },
          "422": {
            "description": "请求体验证失败"
          },
          "500": {
            "description": "服务器内部错误（解析失败、向量索引异常等）"
          }
        }
      }
    },
    "/api/v1/knowledge/preview/{doc_id}": {
      "get": {
        "tags": [
          "知识库"
        ],
        "summary": "文档内容预览",
        "description": "预览文档内容，分页查看切片\n\n- **doc_id**: 文档ID\n- **slice_start**: 起始切片序号\n- **slice_count**: 切片数量",
        "operationId": "get_doc_preview_api_v1_knowledge_preview__doc_id__get",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Doc Id"
            }
          },
          {
            "name": "slice_start",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Slice Start"
            }
          },
          {
            "name": "slice_count",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Slice Count"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/trace/{resource_id}": {
      "get": {
        "tags": [
          "知识库"
        ],
        "summary": "知识溯源查询",
        "description": "根据资源ID反向溯源知识库原始文档\n\n- **resource_id**: 学习资源ID",
        "operationId": "trace_resource_knowledge_api_v1_knowledge_trace__resource_id__get",
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Resource Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/knowledge/docs/{doc_id}/reindex": {
      "post": {
        "tags": [
          "知识库"
        ],
        "summary": "重新索引文档",
        "description": "重新对文档进行切片和向量索引",
        "operationId": "reindex_doc_api_v1_knowledge_docs__doc_id__reindex_post",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Doc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "KnowledgeBatchDeleteRequest": {
        "properties": {
          "doc_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Doc Ids"
          }
        },
        "type": "object",
        "required": [
          "doc_ids"
        ],
        "title": "KnowledgeBatchDeleteRequest",
        "description": "批量删除请求"
      },
      "KnowledgeDocUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "is_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Enabled"
          }
        },
        "type": "object",
        "title": "KnowledgeDocUpdate",
        "description": "更新文档请求"
      },
      "KnowledgeSearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Query",
            "description": "检索查询"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry",
            "description": "行业过滤"
          },
          "doc_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc Id",
            "description": "指定文档ID"
          },
          "top_k": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Top K",
            "description": "返回结果数量",
            "default": 5
          },
          "min_similarity": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Min Similarity",
            "description": "最低相似度",
            "default": 0.6
          },
          "search_type": {
            "type": "string",
            "title": "Search Type",
            "description": "检索类型(vector/keyword/hybrid)",
            "default": "hybrid"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "KnowledgeSearchRequest",
        "description": "知识库检索请求"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  },
...
}